Fix scala-js/scala-js#1287: Reorganize package names and artifacts
