+++ js-envs/src/main/scala/org/scalajs/jsenv/phantomjs/JettyWebsocketManager.scala
@@ -1,4 +1,4 @@
-package scala.scalajs.sbtplugin.env.phantomjs
+package org.scalajs.jsenv.phantomjs
 
 import javax.servlet.http.HttpServletRequest
 
@@ -8,7 +8,7 @@ import org.eclipse.jetty.websocket.{WebSocket, WebSocketHandler}
 import org.eclipse.jetty.util.component.{LifeCycle, AbstractLifeCycle}
 import org.eclipse.jetty.util.log
 
-private[phantomjs] final class JettyWebsocketManager(
+final class JettyWebsocketManager(
     wsListener: WebsocketListener) extends WebsocketManager { thisMgr =>
 
   private[this] var webSocketConn: WebSocket.Connection = null
