+++ js-envs/src/main/scala/org/scalajs/jsenv/phantomjs/WebsocketListener.scala
@@ -1,4 +1,4 @@
-package scala.scalajs.sbtplugin.env.phantomjs
+package org.scalajs.jsenv.phantomjs
 
 private[phantomjs] trait WebsocketListener {
   def onRunning(): Unit
