+++ js-envs/src/main/scala/org/scalajs/jsenv/phantomjs/WebsocketManager.scala
@@ -1,4 +1,4 @@
-package scala.scalajs.sbtplugin.env.phantomjs
+package org.scalajs.jsenv.phantomjs
 
 private[phantomjs] trait WebsocketManager {
   def start(): Unit
