+++ sbt-plugin-test/project/build.sbt
@@ -1,4 +1,4 @@
-addSbtPlugin("org.scala-lang.modules.scalajs" % "scalajs-sbt-plugin" %
-  scala.scalajs.ir.ScalaJSVersions.current)
+addSbtPlugin("org.scala-js" % "sbt-scalajs" %
+  org.scalajs.core.ir.ScalaJSVersions.current)
 
 libraryDependencies += "org.eclipse.jetty" % "jetty-server" % "9.2.3.v20140905"
